requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
distro==1.9.0
eval_type_backport==0.2.0
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
neo4j==5.25.0
neo4j-graphrag==1.0.0
openai==1.51.2
pydantic==2.9.2
pydantic_core==2.23.4
pypdf==4.3.1
python-dotenv==1.0.1
pytz==2024.2
sentence-transformers==3.0.0
sniffio==1.3.1
tqdm==4.66.5
types-mock==5.1.0.20240425
typing_extensions==4.12.2
